+++ components/header/AppHeader.vue
@@ -11,21 +11,25 @@
   <!-- Header of the app -->
   <header
     id="header"
-    class="fixed top-0 right-0 z-50 flex items-center justify-between w-full gap-6 px-6 py-4 transition-all rounded-b-lg shadow-lg justify-self-center h-min bg-primary-light dark:bg-primary-dark md:px-8 scroll-reduce"
+    class="fixed top-0 right-0 z-50 flex flex-col w-full transition-all shadow-lg h-min bg-primary-light dark:bg-primary-dark"
   >
-    <nuxt-icon
-      name="round-list"
-      class="text-3xl cursor-pointer fill-current hover:text-purple-500"
-      :class="{ 'text-purple-700' : menuOpen }"
-      role="button"
-      aria-label="Toggle menu"
-      :aria-expanded="menuOpen.toString()"
-      @click="toggleMenuDropdown"
-    />
-    <div class="flex flex-row items-center gap-3 md:gap-6">
-      <HeaderThemeSwitch aria-label="Switch theme" />
-      <HeaderLanguageSelector aria-label="Select language" />
+    <div class="flex items-center justify-between w-full gap-6 px-6 py-4 md:px-8">
+      <nuxt-icon
+        name="round-list"
+        class="text-3xl cursor-pointer fill-current hover:text-purple-500"
+        :class="{ 'text-purple-700' : menuOpen }"
+        role="button"
+        aria-label="Toggle menu"
+        :aria-expanded="menuOpen.toString()"
+        @click="toggleMenuDropdown"
+      />
+      <div class="flex flex-row items-center gap-3 md:gap-6">
+        <HeaderThemeSwitch />
+        <HeaderLanguageSelector />
+      </div>
     </div>
+    <!-- Progess bar -->
+    <hr class="right-0 w-full border border-purple-500/80 scroll-reduce">
   </header>
 </template>
 
@@ -42,16 +46,13 @@ const toggleMenuDropdown = () => {
 <style scoped>
 @keyframes reduce {
   to {
-    padding-top: 0.5rem;
-    padding-bottom: 0.5rem;
-    width: 206px;
+    width: 1px;
   }
-
 }
 
 .scroll-reduce {
   animation: linear reduce both;
-  animation-range: 0 100vh;
+  animation-range: 0 100%;
   animation-timeline: scroll(root);
 }