+++ components/header/ThemeSwitch.vue
@@ -1,5 +1,5 @@
 <template>
-  <div class="flex items-center cursor-pointer" aria-label="Change color theme" @click="setTheme">
+  <div class="flex items-center cursor-pointer" @click="setTheme">
     <nuxt-icon v-if="themeStore.isLightMode" name="sun" class="text-2xl hover:text-yellow-500" />
     <nuxt-icon v-else name="moon" class="text-2xl hover:text-purple-500" />
   </div>