+++ components/presentation/MyPresentation.vue
@@ -1,7 +1,7 @@
 <template>
   <section
     :id="sectionId"
-    class="flex flex-col min-h-dvh gap-4 p-6 lg:flex-row lg:gap-10 lg:px-32 lg:py-16"
+    class="flex flex-col gap-4 p-6 min-h-dvh lg:flex-row lg:gap-10 lg:px-32 lg:py-16"
     role="region"
     :aria-label="sectionId"
   >
@@ -14,7 +14,7 @@
 </template>
 
 <script setup lang="ts">
-import { type PresentationInfo } from '~/types/types.d'
+import { type PresentationInfo } from '~/lib/types.d'
 
 const props = defineProps<{ sectionData: { [title: string]: PresentationInfo } }>()