+++ components/skills/TechnicalSkillsList.vue
@@ -17,7 +17,7 @@
 </template>
 
 <script setup lang="ts">
-import { type Skill } from '~/types/types.d'
+import { type Skill } from '~/lib/types.d'
 
 const props = defineProps<{ data: Skill }>()